Move allowed_keys to shared helper method
